Refactor: extract search page module into web/js/pages/search.js (#79)
